Do not prompt more than once for the same placeholder
When a placeholder was included inside various boolean (if/else) blocks,
microgen would prompt the user for it's value for each instance and overwrite
the cached answer for the placeholder with the last entry.

Now, microgen properly ignores a placeholder when it's already been answered
and the template will render accroding to the first and only provided answer
for that placeholder.